Add gradle run task, remove old build.gradle.kts
