+++ build.gradle.kts
@@ -12,7 +12,6 @@ plugins {
     java
     kotlin("jvm") version ktVersion
     kotlin("kapt") version ktVersion
-    sciview.base
     sciview.publish
     sciview.sign
     id("org.jetbrains.dokka") version dokkaVersion
@@ -221,6 +220,35 @@ tasks {
                 }
             }
         }
+
+    register<JavaExec>(name = "run") {
+        classpath = sourceSets.main.get().runtimeClasspath
+        if (project.hasProperty("target")) {
+            project.property("target")?.let { target ->
+                classpath = sourceSets.test.get().runtimeClasspath
+
+                println("Target is $target")
+//                if(target.endsWith(".kt")) {
+//                    main = target.substringAfter("kotlin${File.separatorChar}").replace(File.separatorChar, '.').substringBefore(".kt")
+//                } else {
+//                    main = target.substringAfter("java${File.separatorChar}").replace(File.separatorChar, '.').substringBefore(".java")
+//                }
+
+                main = "$target"
+                val props = System.getProperties().filter { (k, _) -> k.toString().startsWith("scenery.") }
+
+                val additionalArgs = System.getenv("SCENERY_JVM_ARGS")
+                allJvmArgs = if (additionalArgs != null) {
+                    allJvmArgs + props.flatMap { (k, v) -> listOf("-D$k=$v") } + additionalArgs
+                } else {
+                    allJvmArgs + props.flatMap { (k, v) -> listOf("-D$k=$v") }
+                }
+
+                println("Will run target $target with classpath $classpath, main=$main")
+                println("JVM arguments passed to target: $allJvmArgs")
+          }
+        }
+    }
 }
 
 val dokkaJavadocJar by tasks.register<Jar>("dokkaJavadocJar") {
